Make Label abstract
